requirements.txt
beautifulsoup4==4.11.1
pandas==1.1.5
Pillow==9.3.0
Plost==0.2.5
requests==2.26.0
streamlit==1.8.1
streamlit_authenticator==0.1.3
